+++ fhircraft/fhir/mapper/engine/exceptions.py
@@ -1,0 +1,16 @@
+class MappingError(Exception):
+    """Base exception for mapping engine errors."""
+
+    pass
+
+
+class ValidationError(MappingError):
+    """Raised when input data validation fails."""
+
+    pass
+
+
+class RuleProcessingError(MappingError):
+    """Raised when rule processing fails."""
+
+    pass
